requirements.txt
chromadb
langchain-ollama
streamlit